Resolve circular dependency
