+++ js/core/renderer_base.js
@@ -4,7 +4,7 @@ var $ = require("jquery");
 var dataUtils = require("./element_data");
 var rendererStrategy = require("./native_renderer_strategy");
 var typeUtils = require("./utils/type");
-var HTMLParser = require("./utils/htmlParser");
+var merge = require("./utils/array").merge;
 var matches = require("./polyfills/matches");
 
 var methods = [
@@ -157,6 +157,71 @@ initRender.prototype.toggleClass = function(className, value) {
     return this;
 };
 
+var isTagName = (/<([a-z][^\/\0>\x20\t\r\n\f]+)/i);
+
+var tagWrappers = {
+    default: {
+        tagsCount: 0,
+        startTags: "",
+        endTags: ""
+    },
+    thead: {
+        tagsCount: 1,
+        startTags: "<table>",
+        endTags: "</table>"
+    },
+    td: {
+        tagsCount: 3,
+        startTags: "<table><tbody><tr>",
+        endTags: "</tr></tbody></table>"
+    },
+    col: {
+        tagsCount: 2,
+        startTags: "<table><colgroup>",
+        endTags: "</colgroup></table>"
+    },
+    tr: {
+        tagsCount: 2,
+        startTags: "<table><tbody>",
+        endTags: "</tbody></table>"
+    },
+};
+
+tagWrappers.tbody = tagWrappers.colgroup = tagWrappers.caption = tagWrappers.tfoot = tagWrappers.thead;
+tagWrappers.th = tagWrappers.td;
+
+var removeScriptTags = function(fragment) {
+    var searchResult = fragment.querySelectorAll("script");
+
+    for(var i = 0; i < searchResult.length; i++) {
+        renderer(searchResult[i]).remove();
+    }
+};
+
+var parseHTML = function(html, context) {
+    if(typeof html !== "string") {
+        return null;
+    }
+
+    var fragment = document.createDocumentFragment();
+    var container = fragment.appendChild(document.createElement("div"));
+
+    var firstRootTag = (isTagName.exec(html) || [ "", "" ])[1].toLowerCase();
+    var tagWrapper = tagWrappers[firstRootTag] || tagWrappers.default;
+
+    container.innerHTML = tagWrapper.startTags + html + tagWrapper.endTags;
+
+    removeScriptTags(fragment);
+
+    for(var i = 0; i < tagWrapper.tagsCount; i++) {
+        container = container.lastChild;
+    }
+
+    var nodes = merge([], container.childNodes);
+
+    return nodes;
+};
+
 initRender.prototype.html = function(value) {
     if(!arguments.length) {
         return this[0].innerHTML;
@@ -174,7 +239,7 @@ initRender.prototype.html = function(value) {
         }
     }
 
-    return this.append(HTMLParser(value));
+    return this.append(parseHTML(value));
 };
 
 var appendElements = function(element, nextSibling) {
